lowguard_core/requirements.txt
fastapi
uvicorn
paho-mqtt
pydantic
onvif-zeep
wsdiscovery
python-multipart
httpx
ffmpeg-python
zeroconf
bcrypt
cryptography
qrcode
pyyaml
